Upgrade to 2.3.4

+++ src/org/hsqldb/lib/ArrayCounter.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2014, The HSQL Development Group
+/* Copyright (c) 2001-2016, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -126,14 +126,14 @@ public class ArrayCounter {
                 return start;
             }
 
-            currentLimit = start + interval < limit ? (int) (start + interval)
+            currentLimit = start + interval < limit ? (start + interval)
                                                     : limit;
         }
     }
 
     /**
      * Helper method to calculate the span of the sub-interval. Simply returns
-     * the cieling of ((limit - start) / segments) and accounts for invalid
+     * the ceiling of ((limit - start) / segments) and accounts for invalid
      * start and limit combinations.
      */
     static int calcInterval(int segments, int start, int limit) {
